Add a threaded interpreter, for situations where you need to ensure service isn't interrupted.
Say you do `while True: pass` in a manhole - you've just trashed your
eventloop, and restarting the process is the only option. With the
ThreadedInteractiveInterpreter, you'll have a thread that thrashes the
CPU, but at least the event loop will keep running.

Of course, this means that redirecting stdout can get pretty ugly, given
we're doing it across threads - the timeout on this at least ensures
that the redirection won't be permanent if something goes wrong in the
thread.