Do not attempt to clean up socket file on shutdown.
Because we were adding the callback and not exploding if there was an
error creating the manhole, any errors that did happen did so silently.

Cleaning up UNIX sockets on shutdown is not really a good practice
anyway, because you can't rely on clean shutdowns. The typical approach
is to clean them up on start up.